Bootstrap the application using standalone APIs
